trade execute: a lot of real world fixes and tweaks
In real world, setting an order is not the end of the process, because
it may not execute in due time because no one is interested in the set
rate anymore. To aleviate the issue, we apply several tweaks to trading
process, specifically:

- we attempt to pick a slightly higher rate than the last bid to get on
top of the order queue (assuming the price is not much higher than
originally planned);

- we reserve less gold for floating pennies to fully use the power of
the strategy for the available funds;

- we try to set buy orders before all sell orders are complete, hoping
there is enough gold funds to fulfill requests (if not, we continue over
and over until success);

- if orders are not fulfilled in 5 minutes, we respin the whole process
hoping new prices and new weights will give us a better result;

- we clear up any open orders before doing the ordering, otherwise funds
are blocked in those orders, which scews the strategy weight
calculation;

- at least Poloniex subtracts fees from the bought alts/gold itself, so
we don't need to accommodate for the fee when setting an order, exchange
will handle it for us.